implement LFA feedback
