+++ .travis.yml
@@ -10,7 +10,7 @@ script:
   - psql -c 'create database stackoverflow;' -U postgres
   - python migration.py
   - coverage run -m pytest
-  - py.test --cov=project project/tests/
+  - py.test --cov=project
 
 after_success:
   - coveralls
