+++ project/__init__.py
@@ -1,12 +1,12 @@
 
 import os
 from flask import Flask
-from project.questions.views import stackoverflow
-from project.models.models import users
+from project.questions.views import questions
+from project.users.views import users
 
 app = Flask(__name__)
 
-app.register_blueprint(stackoverflow, url_prefix='/api/v1')
+app.register_blueprint(questions, url_prefix='/api/v1')
 app.register_blueprint(users, url_prefix='/api/v1')
 
 app_settings = os.getenv(
@@ -14,3 +14,5 @@ app_settings = os.getenv(
     'project.config.DevConfig'
 )
 app.config.from_object(app_settings)
+
+print(app.config.get("DATABASE_HOST"))
