--- project/models/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-
-# configure blueprint
-users = Blueprint('users', __name__)
