+++ tests/test_questions.py
@@ -1,38 +1,10 @@
-import unittest
 import json
-from project import app
-from project.config import create_tables, delete_tables
+import unittest
+
+from . import BaseTest
 
 
-class TestQuestions(unittest.TestCase):
-
-    def setUp(self):
-        app.testing = True
-        self.app = app
-        self.client = self.app.test_client
-
-        # create db tables
-        create_tables()
-
-        # create user
-        self.client().post('api/v1/auth/signup', data=json.dumps(
-            dict(email='test@gmail.com', username='test', password='password')),
-                           content_type='application/json')
-
-        # add question
-        self.client().post('api/v1/questions', data=json.dumps(
-            dict(title='test title', body='some body of quiz')),
-                           content_type='application/json',
-                           headers=dict(token=self.login()))
-
-    def tearDown(self):
-        delete_tables()
-
-    def login(self):
-        res_login = self.client().post('api/v1/auth/login', data=json.dumps(
-            dict(username='test', password='password')),
-                                       content_type='application/json')
-        return json.loads(res_login.data.decode())["token"]
+class TestQuestions(BaseTest):
 
     def test_get_questions(self):
         res = self.client().get('/api/v1/questions', content_type='application/json', headers=dict(token=self.login()))
@@ -83,14 +55,31 @@ class TestQuestions(unittest.TestCase):
 
         assert resp.status_code == 401
         resp_data = json.loads(resp.data.decode())
-        self.assertEqual(resp_data['error'], 'could not generate user id from token')
+        self.assertEqual(resp_data['response'], 'could not generate user id from token')
 
-    # TODO: finish test questions
     def test_delete_question(self):
         self.client().post('api/v1/questions', data=json.dumps(
             dict(title='test title', body='some body of quiz')),
                            content_type='application/json',
                            headers=dict(token=self.login()))
+
+        # add another question
+        self.client().post('api/v1/questions', data=json.dumps(
+            dict(title='test title 1', body='some body of quiz 1')),
+                           content_type='application/json',
+                           headers=dict(token=self.login()))
+
+        res = self.client().get('/api/v1/questions', content_type='application/json',
+                                headers=dict(token=self.login()))
+        resp_data = json.loads(res.data.decode())
+
+        qid = resp_data['questions'][1]['qid']
+
+        resp = self.client().delete('api/v1/questions/' + qid, content_type='application/json',
+                                    headers=dict(token=self.login()))
+        assert resp.status_code == 200
+        resp_data = json.loads(resp.data.decode())
+        self.assertEqual(resp_data['response'], 'question deleted successfully')
 
 
 if __name__ == '__main__':
